requirements.txt
networkx[default]
python_gedcom_2
pygraphviz
pydot
matplotlib
neo4j
unidecode
dateutils